Consolidate Huawei AH100/CH100 handler into HuaweiAhCh100Handler
Addresses review feedback on #1374: merge HuaweiAH100Handler,
HuaweiCH100Handler, HuaweiAhCh100ScaleHandler and the libs
HuaweiAhCh100Protocol object into one HuaweiAhCh100Handler class.
The protocol primitives are now an internal companion object;
HuaweiAhCh100ProtocolTest is retargeted as HuaweiAhCh100HandlerTest.
