--- android_app/app/src/main/java/com/health/openscale/core/bluetooth/scales/HuaweiAH100Handler.kt
@@ -1,30 +1,0 @@
-/*
- * openScale
- * Copyright (C) 2026 openScale contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package com.health.openscale.core.bluetooth.scales
-
-/**
- * Huawei AH100 body-fat scale.
- *
- * Same hardware as the [HuaweiCH100Handler]; the two products only differ
- * in the BLE advertisement name. All wire-protocol logic lives in
- * [HuaweiAhCh100ScaleHandler] and is exercised by `HuaweiAhCh100ProtocolTest`.
- */
-class HuaweiAH100Handler : HuaweiAhCh100ScaleHandler() {
-    override val supportedAdvertName: String = "AH100"
-    override val displayName: String = "Huawei AH100"
-}
